requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
piper-tts>=1.3.0
